pyproject.toml
[project]
name = "diobank"
version = "0.1.0"
description = ""
authors = [
    {name = "Rodrigo Kennes",email = "rodrigo.kennes@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi",
    "uvicorn[standard] (>=0.35.0,<0.36.0)",
    "databases[aiosqlite] (>=0.9.0,<0.10.0)",
    "pyjwt (>=2.10.1,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
